Crate bevy_image

Modules

prelude

Structs

CompressedImageFormats

FileTextureError

An error that occurs when loading a texture from a file.

HdrTextureLoader

Loads HDR textures as Texture assets

HdrTextureLoaderSettings

Image

ImageLoader

Loader for images that can be read by the image crate.

ImageLoaderSettings

ImageSamplerDescriptor

Indicates to an ImageLoader how an Image should be sampled.

Enums

DataFormat

HdrTextureLoaderError

ImageAddressMode

How edges should be handled in texture addressing.

ImageCompareFunction

Comparison function used for depth and stencil operations.

ImageFilterMode

Texel mixing mode when sampling between texels.

ImageFormat

ImageFormatSetting

ImageLoaderError

ImageSampler

Used in Image, this determines what image sampler to use when rendering. The default setting, ImageSampler::Default, will read the sampler from the ImagePlugin at setup. Setting this to ImageSampler::Descriptor will override the global default descriptor for this Image.

ImageSamplerBorderColor

Color variation to use when the sampler addressing mode is ImageAddressMode::ClampToBorder.

ImageType

The type of a raw image buffer.

IntoDynamicImageError

Errors that occur while converting an Image into a DynamicImage

TextureAccessError

An error that occurs when accessing specific pixels in a texture

TextureError

An error that occurs when loading a texture

TranscodeFormat

Constants

SAMPLER_ASSET_INDEX

TEXTURE_ASSET_INDEX

Traits

BevyDefault

TextureFormatPixelInfo

Extends the wgpu TextureFormat with information about the pixel.

Volume

Used to calculate the volume of an item.

Functions

ktx2_buffer_to_image

ktx2_dfd_to_texture_format

ktx2_format_to_texture_format

ktx2_get_texture_format